Move and rename core.py to onxruntime.py
